feat: api 서버 아키텍처 변경
