+++ backend/apps/api/src/models/schemas.py
@@ -1,0 +1,16 @@
+from pydantic import BaseModel
+
+class UploadRequest(BaseModel):
+    filename: str
+    contentType: str
+    size: int
+
+class UploadResponse(BaseModel):
+    presignedUrl: str
+    objectKey: str
+
+class ImageResponse(BaseModel):
+    url: str
+    key: str
+    expiresIn: int
+    download: bool
